+++ src/application/index.ts
@@ -5,3 +5,4 @@ export { BuildService } from './BuildService';
 export { UsageService } from './UsageService';
 export { RevisionQueueService } from './RevisionQueueService';
 export { MotifLedgerService } from './MotifLedgerService';
+export { QueryService } from './QueryService';